Improve structure and JSDoc of shared package
