--- ui/packages/shared/src/core/plugins.ts
@@ -1,5 +1,0 @@
-import type { PluginModule } from "../types/plugin";
-
-export function definePlugin(plugin: PluginModule): PluginModule {
-  return plugin;
-}
