+++ ui/packages/shared/src/index.ts
@@ -1,17 +1,4 @@
-export * from "./core/plugins";
-export * from "./states/attachment-selector";
-export * from "./states/backup";
-export * from "./states/comment";
-export * from "./states/comment-subject-ref";
-export * from "./states/editor";
-export * from "./states/entity";
-export * from "./states/operation";
-export * from "./states/plugin-installation-tabs";
-export * from "./states/plugin-tab";
-export * from "./states/theme-list-tabs";
-export * from "./states/user-tab";
+export * from "./plugin";
 export * from "./stores";
-export * from "./types/menus";
-export * from "./types/plugin";
-export * from "./types/widget";
+export * from "./types";
 export * from "./utils";
