+++ ui/packages/shared/src/plugin/types/index.ts
@@ -1,0 +1,13 @@
+export * from "./attachment-selector";
+export * from "./backup";
+export * from "./comment";
+export * from "./dashboard-widget";
+export * from "./editor-provider";
+export * from "./list-entity-field";
+export * from "./list-operation";
+export * from "./plugin-installation-tab";
+export * from "./plugin-tab";
+export * from "./theme-list-tab";
+export * from "./ui-plugin-entry";
+export * from "./ui-plugin-module";
+export * from "./user-tab";
