--- ui/packages/shared/src/states/backup.ts
@@ -1,8 +1,0 @@
-import type { Component, Raw } from "vue";
-
-export interface BackupTab {
-  id: string;
-  label: string;
-  component: Raw<Component>;
-  permissions?: string[];
-}
