--- ui/packages/shared/src/states/comment.ts
@@ -1,9 +1,0 @@
-import type { Component, Raw } from "vue";
-
-export interface CommentEditorProvider {
-  component: Raw<Component>;
-}
-
-export interface CommentContentProvider {
-  component: Raw<Component>;
-}
