--- ui/packages/shared/src/states/editor.ts
@@ -1,9 +1,0 @@
-import type { Component } from "vue";
-
-export interface EditorProvider {
-  name: string;
-  displayName: string;
-  logo?: string;
-  component: Component;
-  rawType: string;
-}
