--- ui/packages/shared/src/states/entity.ts
@@ -1,10 +1,0 @@
-import type { Component, Raw } from "vue";
-
-export interface EntityFieldItem {
-  priority: number;
-  position: "start" | "end";
-  component: Raw<Component>;
-  props?: Record<string, unknown>;
-  permissions?: string[];
-  hidden?: boolean;
-}
