--- ui/packages/shared/src/states/operation.ts
@@ -1,12 +1,0 @@
-import type { Component, Raw } from "vue";
-
-export interface OperationItem<T> {
-  priority: number;
-  component: Raw<Component>;
-  props?: Record<string, unknown>;
-  action?: (item?: T) => void;
-  label?: string;
-  hidden?: boolean;
-  permissions?: string[];
-  children?: OperationItem<T>[];
-}
