--- ui/packages/shared/src/states/plugin-tab.ts
@@ -1,8 +1,0 @@
-import type { Component, Raw } from "vue";
-
-export interface PluginTab {
-  id: string;
-  label: string;
-  component: Raw<Component>;
-  permissions?: string[];
-}
