--- ui/packages/shared/src/states/theme-list-tabs.ts
@@ -1,10 +1,0 @@
-import type { Component, Raw } from "vue";
-
-export interface ThemeListTab {
-  id: string;
-  label: string;
-  component: Raw<Component>;
-  props?: Record<string, unknown>;
-  permissions?: string[];
-  priority: number;
-}
