--- ui/packages/shared/src/types/plugin.ts
@@ -1,135 +1,0 @@
-import type { BackupTab } from "@/states/backup";
-import type {
-  CommentContentProvider,
-  CommentEditorProvider,
-} from "@/states/comment";
-import type { CommentSubjectRefProvider } from "@/states/comment-subject-ref";
-import type { EntityFieldItem } from "@/states/entity";
-import type { OperationItem } from "@/states/operation";
-import type { PluginInstallationTab } from "@/states/plugin-installation-tabs";
-import type { ThemeListTab } from "@/states/theme-list-tabs";
-import type { UserProfileTab, UserTab } from "@/states/user-tab";
-import type {
-  Attachment,
-  Backup,
-  ListedComment,
-  ListedPost,
-  ListedReply,
-  ListedSinglePage,
-  Plugin,
-  Theme,
-} from "@halo-dev/api-client";
-import type { AnyExtension } from "@halo-dev/richtext-editor";
-import type { Component, Ref } from "vue";
-import type { RouteRecordName, RouteRecordRaw } from "vue-router";
-import type {
-  DashboardWidgetDefinition,
-  DashboardWidgetQuickActionItem,
-  EditorProvider,
-  PluginTab,
-} from "..";
-import type { AttachmentSelectProvider } from "../states/attachment-selector";
-
-export interface RouteRecordAppend {
-  parentName: NonNullable<RouteRecordName>;
-  route: RouteRecordRaw;
-}
-
-export interface ExtensionPoint {
-  "attachment:selector:create"?: () =>
-    | AttachmentSelectProvider[]
-    | Promise<AttachmentSelectProvider[]>;
-
-  "editor:create"?: () => EditorProvider[] | Promise<EditorProvider[]>;
-
-  "plugin:self:tabs:create"?: () => PluginTab[] | Promise<PluginTab[]>;
-
-  "default:editor:extension:create"?: () =>
-    | AnyExtension[]
-    | Promise<AnyExtension[]>;
-
-  "comment:subject-ref:create"?: () => CommentSubjectRefProvider[];
-
-  "comment:editor:replace"?: () =>
-    | CommentEditorProvider
-    | Promise<CommentEditorProvider>;
-
-  "comment:list-item:content:replace"?: () =>
-    | CommentContentProvider
-    | Promise<CommentContentProvider>;
-
-  "backup:tabs:create"?: () => BackupTab[] | Promise<BackupTab[]>;
-
-  "plugin:installation:tabs:create"?: () =>
-    | PluginInstallationTab[]
-    | Promise<PluginInstallationTab[]>;
-
-  "post:list-item:operation:create"?: (
-    post: Ref<ListedPost>
-  ) => OperationItem<ListedPost>[];
-
-  "single-page:list-item:operation:create"?: (
-    singlePage: Ref<ListedSinglePage>
-  ) => OperationItem<ListedSinglePage>[];
-
-  "comment:list-item:operation:create"?: (
-    comment: Ref<ListedComment>
-  ) => OperationItem<ListedComment>[];
-
-  "reply:list-item:operation:create"?: (
-    reply: Ref<ListedReply>
-  ) => OperationItem<ListedReply>[];
-
-  "plugin:list-item:operation:create"?: (
-    plugin: Ref<Plugin>
-  ) => OperationItem<Plugin>[];
-
-  "backup:list-item:operation:create"?: (
-    backup: Ref<Backup>
-  ) => OperationItem<Backup>[];
-
-  "attachment:list-item:operation:create"?: (
-    attachment: Ref<Attachment>
-  ) => OperationItem<Attachment>[];
-
-  "plugin:list-item:field:create"?: (plugin: Ref<Plugin>) => EntityFieldItem[];
-
-  "post:list-item:field:create"?: (post: Ref<ListedPost>) => EntityFieldItem[];
-
-  "single-page:list-item:field:create"?: (
-    singlePage: Ref<ListedSinglePage>
-  ) => EntityFieldItem[];
-
-  "theme:list:tabs:create"?: () => ThemeListTab[] | Promise<ThemeListTab[]>;
-
-  "theme:list-item:operation:create"?: (
-    theme: Ref<Theme>
-  ) => OperationItem<Theme>[];
-
-  "user:detail:tabs:create"?: () => UserTab[] | Promise<UserTab[]>;
-
-  "uc:user:profile:tabs:create"?: () =>
-    | UserProfileTab[]
-    | Promise<UserProfileTab[]>;
-
-  "console:dashboard:widgets:create"?: () =>
-    | DashboardWidgetDefinition[]
-    | Promise<DashboardWidgetDefinition[]>;
-
-  "console:dashboard:widgets:internal:quick-action:item:create"?: () =>
-    | DashboardWidgetQuickActionItem[]
-    | Promise<DashboardWidgetQuickActionItem[]>;
-}
-
-export interface PluginModule {
-  /**
-   * These components will be registered when plugin is activated.
-   */
-  components?: Record<string, Component>;
-
-  routes?: RouteRecordRaw[] | RouteRecordAppend[];
-
-  ucRoutes?: RouteRecordRaw[] | RouteRecordAppend[];
-
-  extensionPoints?: ExtensionPoint;
-}
